+++ src/components/Dialog.vue
@@ -1,6 +1,6 @@
 <template>
   <div
-    class="fixed w-full px-3 md:px-0 h-full flex flex-col justify-center items-center bg-black bg-opacity-60 z-50 top-0 left-0"
+    class="fixed w-full px-3 md:px-0 h-full flex flex-col justify-center items-center bg-black bg-opacity-60 z-100 top-0 left-0"
     :class="props.classname"
   >
     <slot></slot>
