+++ src/pages/Blog.vue
@@ -98,7 +98,7 @@
     <BytebaseBanner v-if="!state.isLoading" class="mt-32 mb-8"></BytebaseBanner>
     <Footer />
 
-    <BytebaseFixedBanner />
+    <SponsorBanner />
   </div>
 </template>
 
@@ -110,7 +110,7 @@ import utils from "../../common/utils";
 import Footer from "../components/Footer.vue";
 import Header from "../components/Header.vue";
 import BytebaseBanner from "../components/BytebaseBanner.vue";
-import BytebaseFixedBanner from "../components/BytebaseFixedBanner.vue";
+import SponsorBanner from "../components/SponsorBanner.vue";
 
 interface State {
   isLoading: boolean;
