+++ src/pages/BlogList.vue
@@ -126,7 +126,7 @@
     <BytebaseBanner v-if="!state.isLoading" class="mb-8"></BytebaseBanner>
     <Footer />
 
-    <BytebaseFixedBanner />
+    <SponsorBanner />
   </div>
 </template>
 
@@ -136,7 +136,7 @@ import utils from "../../common/utils";
 import Footer from "../components/Footer.vue";
 import Header from "../components/Header.vue";
 import BytebaseBanner from "../components/BytebaseBanner.vue";
-import BytebaseFixedBanner from "../components/BytebaseFixedBanner.vue";
+import SponsorBanner from "../components/SponsorBanner.vue";
 
 interface State {
   isLoading: boolean;
